refactored first plugin
