+++ src-tauri/cabr2_search/Cargo.toml
@@ -23,9 +23,8 @@ thiserror = "1.0.24"
 ureq = { version = "2.1.0", features = ["tls", "json"], optional = true }
 
 [features]
-handler = []
 gestis = [ "roxmltree", "ureq" ]
-tauri_plugin = [ "tauri", "handler" ]
+tauri_plugin = [ "tauri" ]
 default = []
 
 [[bin]]
